Fixes from testing
